move swagger modifications to a dedicated annotation and rename the customizers
