+++ src/main/java/it/aboutbits/springboot/toolbox/boot/type/swagger/CustomTypeModelConverter.java
@@ -6,13 +6,14 @@ import io.swagger.v3.core.converter.ModelConverterContext;
 import io.swagger.v3.oas.models.media.Schema;
 import it.aboutbits.springboot.toolbox.persistence.identity.EntityId;
 import it.aboutbits.springboot.toolbox.reflection.util.RecordReflectionUtil;
+import it.aboutbits.springboot.toolbox.type.CustomType;
 import org.springframework.stereotype.Component;
 
 import java.math.BigDecimal;
 import java.util.Iterator;
 
 @Component
-public class IdentityModelConverter implements ModelConverter {
+public class CustomTypeModelConverter implements ModelConverter {
 
     @Override
     public Schema<?> resolve(
@@ -23,9 +24,9 @@ public class IdentityModelConverter implements ModelConverter {
 
         var type = annotatedType.getType();
 
-        if (type instanceof Class<?> clazz && EntityId.class.isAssignableFrom(clazz)) {
+        if (type instanceof Class<?> clazz && CustomType.class.isAssignableFrom(clazz) && !EntityId.class.isAssignableFrom(
+                clazz)) {
             var constructor = RecordReflectionUtil.getCanonicalConstructor(clazz);
-
             var wrappedType = constructor.getParameters()[0].getType();
 
             if (Short.class.isAssignableFrom(wrappedType)) {
